+++ CHANGELOG.rst
@@ -26,13 +26,35 @@ Added
   The filter bank is implemented in ``pyroomacoustics.acoustics.AntoniOctaveFilterBank``.
 
 
+`0.8.3`_ - 2024-12-08
+---------------------
+
+Bugfix
+~~~~~~
+
+- Fixes issue #382: When providing a ``MicrophoneArray`` object with
+  directivity to ``Room.add_microphone_array``, the directivity was dropped
+  from the object.
+
+- Fixes issues #381: When creating a room with from_corners with multi-band
+  material, and then making it 3D with ``extrude`` with a single band material
+  for the floor and ceiling, an error would occur.
+  After the fix, the materials for floor and ceiling are automatically extended
+  to multi-band, as expected.
+
+- Fixes issue #380: Caused by the attribute ``cartesian`` of ``GridSphere`` not
+  being set properly when the grid is only initialized with a number of points.
+
+- Fixes issue #355: Makes the MicrophoneArray class more bug-proof and adds
+  some tests.
+
 `0.8.2`_ - 2024-11-06
 ---------------------
 
 Changed
 ~~~~~~~
 
-- Makes the ``pyroomacoustics.utilities.resample`` backend is made configurable
+- Makes the ``pyroomacoustics.utilities.resample`` backend configurable
   to avoid ``soxr`` dependency. The resample backend is configurable to
   ``soxr``, ``samplerate``, if these packages are available, and otherwise
   falls back to ``scipy.signal.resample_poly``.
@@ -688,7 +710,8 @@ Changed
    ``pyroomacoustics.datasets.timit``
 
 
-.. _Unreleased: https://github.com/LCAV/pyroomacoustics/compare/v0.8.2...master
+.. _Unreleased: https://github.com/LCAV/pyroomacoustics/compare/v0.8.3...master
+.. _0.8.3: https://github.com/LCAV/pyroomacoustics/compare/v0.8.2...v0.8.3
 .. _0.8.2: https://github.com/LCAV/pyroomacoustics/compare/v0.8.1...v0.8.2
 .. _0.8.1: https://github.com/LCAV/pyroomacoustics/compare/v0.8.0...v0.8.1
 .. _0.8.0: https://github.com/LCAV/pyroomacoustics/compare/v0.7.7...v0.8.0